Move old packs bundling code to legacy status
