+++ multipacks-engine/src/main/java/multipacks/bundling/legacy/BundleResult.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package multipacks.bundling;
+package multipacks.bundling.legacy;
 
 import java.util.HashMap;
 import java.util.function.Supplier;
