+++ multipacks-engine/src/main/java/multipacks/bundling/legacy/PackBundler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package multipacks.bundling;
+package multipacks.bundling.legacy;
 
 import java.io.File;
 import java.io.IOException;
