+++ multipacks-engine/src/main/java/multipacks/bundling/legacy/PackagingFailException.java
@@ -13,8 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package multipacks.bundling;
+package multipacks.bundling.legacy;
 
-public enum BundleIgnore {
-	LICENSES;
+public class PackagingFailException extends RuntimeException {
+	private static final long serialVersionUID = 2251765351034996381L;
+
+	public PackagingFailException(String message) {
+		super(message);
+	}
 }
